Refactor IDataEnumerable interface to be more like traditional IEnumerables.
This makes the code more stateful and complicated, but it fixes a
problem with the old interface which made it impossible to implement a
.Take() or .TakeWhile() method on IDataEnumerables, because there was
no way to force finally blocks to run without either throwing an
exception or iterating to the very, very end of the sequence.

This problem is inherently stateful, so an interface that appears
stateless wouldn't be a good fit, plus this makes more sense to people
who understand IEnumerable and IEnumerator.